requirements.txt
streamlit
openai
faiss-cpu
transformers
torch
scikit-learn